backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-huggingface
huggingface_hub
langgraph
chromadb
pandas
pydantic
python-dotenv
tiktoken
pytest
